Resolve conflicts:
- Take main's constraint_system module split and other tree-level moves;
  the PR's dedup lives in frontend cse.rs and is unaffected.
- Regenerate all 12 example snapshots from the merged code so gate counts
  reflect main's current circuit graphs with this PR's gate dedup applied.
  Hand-picking either side was incorrect: "ours" kept stale pre-optimization
  gate counts, "theirs" dropped the dedup effect on AND-constraint and
  value-index counts.

+++ crates/verifier/src/protocols/shift/monster.rs
@@ -194,7 +194,7 @@ fn evaluate_matrices<F: BinaryField, E: FieldOps<Scalar = F> + From<F>>(
 						ShiftVariant::Sra32 => 6,
 						ShiftVariant::Rotr32 => 7,
 					};
-					evals[shift_id][*amount] +=
+					evals[shift_id][*amount as usize] +=
 						constraint_eval.clone() * &r_y_tensor[value_index.0 as usize];
 				}
 			}
